requirements.txt
asgiref==3.8.1
Django==5.1.5
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
ruff
psycopg2-binary
django-ulid
djangorestframework
markdown       
django-filter
djangorestframework-simplejwt
pillow
django-debug-toolbar
jsonschema
redis
django-redis
celery
flower
locust
ipython